feature: support reissue for auth-required node.
